+++ src/components/ProgramCard/ProgramCard.stories.tsx
@@ -1,0 +1,82 @@
+import React from 'react';
+// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
+import { Story, Meta } from '@storybook/react/types-6-0';
+import { colorOptions } from '../../prebuiltTailwindTheme';
+import { ProgramCard, ProgramCardProps } from '..';
+import { gradientFromClassNameMap, gradientToClassNameMap, gradientViaClassNameMap } from '../../storybookMappers';
+
+export default {
+  title: 'Components/ProgramCard',
+  component: ProgramCard,
+  argTypes: {
+    color: {
+      defaultValue: 'pui-primary',
+      control: {
+        type: 'select',
+        options: colorOptions
+      }
+    },
+    className: {
+      control: {
+        type: 'text'
+      }
+    },
+    imageUrl: {
+      defaultValue: 'https://fakeimg.pl/370x110/',
+      control: {
+        type: 'text'
+      }
+    },
+    title: {
+      defaultValue: 'This is a title',
+      control: {
+        type: 'text'
+      }
+    },
+    percentage: {
+      defaultValue: '50',
+      control: {
+        type: 'text'
+      }
+    },
+    buttonpercentage: {
+      defaultValue: 'View more CTA',
+      control: {
+        type: 'text'
+      }
+    },
+    gradientInitialColor: {
+      defaultValue: 'from-pui-primary',
+      control: {
+        type: 'select',
+        options: gradientFromClassNameMap
+      }
+    },
+    gradientFinalColor: {
+      defaultValue: 'to-pui-secondary',
+      control: {
+        type: 'select',
+        options: gradientToClassNameMap
+      }
+    },
+    gradientMiddleColor: {
+      defaultValue: 'to-pui-current',
+      control: {
+        type: 'select',
+        options: gradientViaClassNameMap
+      }
+    }
+  }
+} as Meta;
+
+/**
+ * A story that displays a list connector
+ *
+ * @param props the story props
+ * @param props.color the color property set on controls
+ * @param props.className the classes for element
+ * @param props.imageUrl the image url set on controls
+ */
+const Template: Story<ProgramCardProps & { color?: string }> = ({ ...props }) => <ProgramCard {...props} />;
+
+export const Default = Template.bind({});