+++ CHANGELOG.md
@@ -25,7 +25,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-## [3.3.0]
+## [3.4.0] 2022-01-10
+
+### Added
+
+- Add percentage cicles component
+- Add percentage list connector component
+- Add percentage card Banners component
+- Add percentage cirlces component
+- Add images with gradient border component
+- Add expert cards
+- Add Event cards
+- Add Program cards
+- Add Resources cards
+
+## [3.3.0] 2021-12-08
 
 ### Fixed